+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Upgrade Selenium.WebDriver package reference to v4.49.0.
+- In `AtataSessionBuilder<TSession, TBuilder>` rename because of typo property `DynamicConfiguratonActions` to `DynamicConfigurationActions`.
 
 ## [4.0.1] - 2026-09-09